+++ bin/routes.dart
@@ -11,18 +11,19 @@ String library_path = library_path_default;
 
 create(String name, String path,
     {String dartTemplate,
-    String htmlTemplate,
-    String cssTemplate,
+    String htmlTemplate: element.htmldefault,
+    String cssTemplate: element.cssdefault,
     bool isDefault: false,
     bool isAbstract,
     String redirectTo,
-    String parent}) async {
+    String parent,
+    bool autonotify: true}) async {
   name = toSnakeCase(name);
   if (dartTemplate == null) {
     dartTemplate = routeDartTemplate(name, "${name}Route", path, isDefault,
-        isAbstract: isAbstract, parent: parent, redirectTo: redirectTo);
+        isAbstract: isAbstract, parent: parent, redirectTo: redirectTo, autonotify: autonotify);
   }
-  htmlTemplate = element.elementHtmlTemplate("${name}-route", htmlTemplate ?? "");
+  htmlTemplate = element.elementHtmlTemplate("${name}-route", htmlTemplate, cssTemplate);
 
   await writeInDartFile(
       "lib/${toSnakeCase(library_path)}/${toSnakeCase(name)}/${toSnakeCase(name)}.dart",
@@ -45,7 +46,7 @@ _notEmptyButNull(String value) {
 }
 
 routeDartTemplate(String name, String routeName, String path, bool isDefault,
-        {bool isAbstract: false, String redirectTo: "", String parent: ""}) =>
+        {bool isAbstract: false, String redirectTo: "", String parent: "", bool autonotify: true}) =>
     '''
       @HtmlImport("${toSnakeCase(name)}.html")
       library route_elements.${toSnakeCase(name)};
@@ -54,8 +55,9 @@ routeDartTemplate(String name, String routeName, String path, bool isDefault,
       @PolyceRoute("${toCamelCase(routeName)}", "$path", isDefault: $isDefault,
       isAbstract: $isAbstract, parent: ${_notEmptyButNull(toCamelCase(parent))}, redirectTo: ${_notEmptyButNull(toCamelCase(redirectTo))})
       @PolymerRegister("${toLispCase(name)}")
-      class ${toCamelCase(name)} extends PolymerElement with AutonotifyBehavior, Observable, PolymerAppRouteBehavior {
+      class ${toCamelCase(name)} extends PolymerElement ${ autonotify ? "with AutonotifyBehavior, Observable, PolymerAppRouteBehavior" : "" } {
       ${toCamelCase(name)}.created() : super.created();
+      ${element.propertyTemplate("field", String, autonotify)}
       /// Called when an instance of ${toLispCase(name)} is inserted into the DOM.
       attached() {
       super.attached();