+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+#changelog
 
 ## 0.2.0
 
@@ -11,3 +11,18 @@
         * `encode` method
         * factory constructor `decode`
     - Config file is now `polyce.settings.yaml`
+
+# 0.1.5
+
+- Fix service generation
+
+# 0.1.4
+
+- Fix init service
+- No css file on element generation
+- default html content in element
+- autonotify behavior can be desactivate
+
+# <= 0.1.3
+
+First version